requirements.txt
bitstring~=3.1.7
pyserial~=3.4
PyQt5~=5.14.2